# src/backend/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Referência de API da loja virtual Urban Sneakers",
    "description": "Essa API tem como objetivo a manipulação de dados e realização de operações da loja Urban Sneakers.",
    "contact": {
      "email": "albertojr.neto@gmail.com"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/features": {
      "post": {
        "summary": "Cadastro de novas características",
        "description": "Rota responsável por cadastrar uma nova característica de produto.",
        "tags": ["Features"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              },
              "examples": {
                "feature": {
                  "value": {
                    "is_multiple": true,
                    "name": "Material"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "ER_DUP_ENTRY: Duplicate entry"
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Feature"
                    },
                    {
                      "$ref": "#/components/schemas/ResourceUrl"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Busca de todas as características",
        "description": "Rota responsável por buscar todas as características de produto.",
        "tags": ["Features"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                },
                "examples": {
                  "feature": {
                    "value": [
                      {
                        "id": 0,
                        "is_multiple": true,
                        "name": "Material"
                      },
                      {
                        "id": 1,
                        "is_multiple": false,
                        "name": "Cor"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/features/{id}": {
      "get": {
        "summary": "Busca uma característica por id",
        "description": "Rota responsável por buscar uma característica de produto.",
        "tags": ["Features"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da característica",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                },
                "examples": {
                  "feature": {
                    "value": [
                      {
                        "id": 0,
                        "is_multiple": true,
                        "name": "Material"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Altera uma característica",
        "description": "Rota responsável por alterar  os valores de uma característica de produto.",
        "tags": ["Features"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              },
              "examples": {
                "feature": {
                  "value": {
                    "is_multiple": true,
                    "name": "Material"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da característica",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "ER_DUP_ENTRY: Duplicate entry"
          },
          "200": {
            "description": "OK. Retorna o id da característica alterada e apenas os valores que foram alterados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                },
                "examples": {
                  "feature": {
                    "value": {
                      "id": 0,
                      "name": "Material"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Exclui uma característica",
        "description": "Rota responsável por excluir uma característica de produto.",
        "tags": ["Features"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id da característica",
            "required": true
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Feature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "is_multiple": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ResourceUrl": {
        "type": "object",
        "properties": {
          "resourceUrl": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}